Only create the __complete command when called
The go tests were failing before this commit, because the new __complete
command was the first subcommand of a test.  This showed that adding
this command could impact Cobra programs that only had a single command
(the root command).  By only creating the __complete command if it is
currently being called, allows all other commands to be unaffected.  It
also fixes the failing go tests.

Signed-off-by: Marc Khouzam <marc.khouzam@gmail.com>